+++ src/css/SplashTitle/SplashTitle.css
