+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./css/index.css";
+import "./css/index/index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
